requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
mixins==0.1.4
pytz==2021.1
sentry-sdk==1.0.0
sqlparse==0.4.1
urllib3==1.26.4
